implement a tx capacity checker or throw error
