+++ apps/account-worker/src/transaction_publisher/publisher.service.ts
@@ -67,7 +67,6 @@ export class TransactionPublisherService extends BaseConsumer implements OnAppli
     try {
       // Check capacity first; if out of capacity, send job back to queue
       if (!(await this.capacityCheckerService.checkForSufficientCapacity())) {
-        job.moveToDelayed(Date.now(), job.token); // fake delay, we just want to avoid processing the current job if we're out of capacity
         throw new DelayedError();
       }
       this.logger.log(`Processing job ${job.id} of type ${job.name}.`);
@@ -138,10 +137,11 @@ export class TransactionPublisherService extends BaseConsumer implements OnAppli
       obj[txHash] = JSON.stringify(status);
       this.cacheManager.hset(TXN_WATCH_LIST_KEY, obj);
     } catch (error: any) {
-      if (!(error instanceof DelayedError)) {
+      if (error instanceof DelayedError) {
+        job.moveToDelayed(Date.now(), job.token);
+      } else {
         this.logger.error('Unknown error encountered: ', error, error?.stack);
       }
-
       throw error;
     }
   }
